+++ src/markus_exam_matcher/__init__.py
@@ -1,0 +1,3 @@
+# TODO: Make module for getting initial contours. WHen done
+# TODO: Update description of box_detection.py to point user to
+# TODO: this module.
