+++ src/markus_exam_matcher/_cnn/emnist_cnn.pt
