+++ src/markus_exam_matcher/_cnn/mnist_cnn.pt
