+++ src/markus_exam_matcher/image_processing/image_transformations.py
